+++ src/builtin/logical.rs
@@ -9,7 +9,7 @@ use crate::{
 };
 built_in! {self, func, scope, logical;
   and => {"and", COMMON, AtLeast(2), {
-    self.mov_bool(Rax, func, scope)?;
+    self.take_bool(Rax, func, scope)?;
     for _ in 1..func.len {
     let boolean = take_arg!(self, func, "Bool", Json::Bool(x) => x).0;
       match boolean {
@@ -35,7 +35,7 @@ built_in! {self, func, scope, logical;
     }
   }},
   or => {"or", COMMON, AtLeast(2), {
-    self.mov_bool(Rax, func, scope)?;
+    self.take_bool(Rax, func, scope)?;
     for _ in 1..func.len {
     let boolean = take_arg!(self, func, "Bool", Json::Bool(x) => x).0;
       match boolean {
@@ -50,7 +50,7 @@ built_in! {self, func, scope, logical;
     scope.mov_tmp_bool(Rax)
   }},
   xor => {"xor", COMMON, AtLeast(2), {
-    self.mov_bool(Rax, func, scope)?;
+    self.take_bool(Rax, func, scope)?;
     for _ in 1..func.len {
     let boolean = take_arg!(self, func, "Bool", Json::Bool(x) => x).0;
       match boolean {